// apps/console/src/feature/experience/model/MallDraftType.ts
import type { MallCreateDraft, MallUpdateDraft, ThemePresetId } from './Mall';

export interface MallOpeningDraft {
  readonly parentId: string;
  readonly name: string;
  readonly code: string;
  readonly publicSlug: string;
  readonly brandName: string;
  readonly domainMode: MallCreateDraft['domain']['mode'];
  readonly customDomain: string;
  readonly ownerMembershipId: string;
  readonly timezone: string;
  readonly currency: string;
  readonly themePreset: ThemePresetId;
  readonly primaryColor: string;
  readonly accentColor: string;
  readonly logoObjectRef: string;
  readonly faviconObjectRef: string;
  readonly subjectType: MallCreateDraft['opening']['subject']['type'];
  readonly companyName: string;
  readonly creditCode: string;
  readonly legalRepresentative: string;
  readonly contactName: string;
  readonly contactMobile: string;
  readonly licenseObjectRef: string;
  readonly storeType: MallCreateDraft['opening']['business']['storeType'];
  readonly primaryCategory: string;
  readonly businessMode: MallCreateDraft['opening']['business']['mode'];
  readonly businessRegion: string;
  readonly businessAddress: string;
  readonly servicePhone: string;
  readonly certificateMode: MallCreateDraft['opening']['certificateMode'];
  readonly certificateObjectRef: string;
  readonly miniProgramMode: MallCreateDraft['opening']['channels']['miniProgramMode'];
  readonly miniProgramAppId: string;
  readonly miniProgramOriginalId: string;
  readonly officialAccountMode: MallCreateDraft['opening']['channels']['officialAccountMode'];
  readonly officialAccountAppId: string;
  readonly videoChannelId: string;
  readonly paymentPlan: MallCreateDraft['opening']['payment']['plan'];
  readonly wechatMerchantId: string;
  readonly deliveryMode: MallCreateDraft['opening']['fulfillment']['deliveryMode'];
  readonly warehouseRegion: string;
  readonly returnContact: string;
  readonly returnAddress: string;
  readonly invoiceMode: MallCreateDraft['opening']['invoiceMode'];
  readonly notificationContact: string;
  readonly status: MallUpdateDraft['status'];
}